refactor: move tz-utils.js into lib/
Consistent with lib/date-utils.js; tz-utils is internal-only (not exported
through node-ical.js, not mentioned in README or .d.ts) so this is not a
breaking change. Update all internal require paths accordingly.
