+++ lib/date-utils.js
@@ -4,7 +4,7 @@
 // Load Temporal polyfill if not natively available
 const Temporal = globalThis.Temporal || require('temporal-polyfill').Temporal;
 
-const tzUtil = require('../tz-utils.js');
+const tzUtil = require('./tz-utils.js');
 
 /**
  * Construct a date-only key (YYYY-MM-DD) from a Date object.
