+++ lib/tz-utils.js
@@ -3,7 +3,7 @@
 
 // Load Temporal polyfill if not natively available (mirrors ical.js)
 const Temporal = globalThis.Temporal || require('temporal-polyfill').Temporal;
-const windowsZones = require('./windowsZones.json');
+const windowsZones = require('../windowsZones.json');
 
 // Ensure polyfill is globally available for downstream modules
 globalThis.Temporal ??= Temporal;
@@ -476,7 +476,6 @@ function linkAlias(arg1, arg2) {
 // Memoize VTIMEZONE→IANA lookups keyed by "stdOffset|dstOffset|year"
 const vtimezoneIanaCache = new Map();
 
-/**
 /**
  * Pick the STANDARD or DAYLIGHT sub-component that applies to a given reference year.
  * A VTIMEZONE may carry multiple historic observance blocks (e.g. the US rule changed in 2007).
